Minor fixes

+++ data/src/main/java/com/canopas/yourspace/data/repository/SpaceRepository.kt
@@ -115,8 +115,8 @@ class SpaceRepository @Inject constructor(
         val spaceId = currentSpaceId
 
         if (spaceId.isEmpty()) {
-            val userId = authService.currentUser?.id ?: ""
-            return getUserSpaces(userId).firstOrNull()?.sortedBy { it.created_at }?.firstOrNull()
+            val userId = authService.currentUser?.id
+            return userId?.let { getUserSpaces(it).firstOrNull()?.sortedBy { it.created_at }?.firstOrNull() }
         }
         return getSpace(spaceId)
     }
@@ -185,8 +185,8 @@ class SpaceRepository @Inject constructor(
     }
 
     suspend fun deleteUserSpaces() {
-        val userId = authService.currentUser?.id ?: ""
-        val allSpace = getUserSpaces(userId).firstOrNull() ?: emptyList()
+        val userId = authService.currentUser?.id
+        val allSpace = userId?.let { getUserSpaces(it).firstOrNull() } ?: emptyList()
         val ownSpace = allSpace.filter { it.admin_id == userId }
         val joinedSpace = allSpace.filter { it.admin_id != userId }
 
@@ -195,19 +195,21 @@ class SpaceRepository @Inject constructor(
         }
 
         joinedSpace.forEach { space ->
-            spaceService.removeUserFromSpace(space.id, userId)
+            if (userId != null) {
+                spaceService.removeUserFromSpace(space.id, userId)
+            }
         }
     }
 
     suspend fun deleteSpace(spaceId: String) {
         invitationService.deleteInvitations(spaceId)
         spaceService.deleteSpace(spaceId)
-        val userId = authService.currentUser?.id ?: ""
+        val userId = authService.currentUser?.id
         currentSpaceId =
-            getUserSpaces(userId).firstOrNull()?.sortedBy { it.created_at }?.firstOrNull()?.id
+            userId?.let { getUserSpaces(it).firstOrNull()?.sortedBy { it.created_at }?.firstOrNull()?.id }
                 ?: ""
 
-        val user = userService.getUser(userId)
+        val user = userId?.let { userService.getUser(it) }
         val updatedSpaceIds = user?.space_ids?.toMutableList()?.apply {
             remove(spaceId)
         } ?: return
@@ -218,9 +220,11 @@ class SpaceRepository @Inject constructor(
     }
 
     suspend fun leaveSpace(spaceId: String) {
-        val userId = authService.currentUser?.id ?: ""
-        spaceService.removeUserFromSpace(spaceId, userId)
-        val user = userService.getUser(userId)
+        val userId = authService.currentUser?.id
+        if (userId != null) {
+            spaceService.removeUserFromSpace(spaceId, userId)
+        }
+        val user = userId?.let { userService.getUser(it) }
         val updatedSpaceIds = user?.space_ids?.toMutableList()?.apply {
             remove(spaceId)
         } ?: return
